# openapi.yaml
openapi: 3.0.3
info:
  title: Percolator API
  version: 0.1.0
  description: |
    Percolator is a high-performance perpetual futures trading engine built on Solana.
    This API provides read access to market data, trades, funding rates, and platform statistics.
    
    ## Features
    - Real-time market data and pricing
    - Trade history and volume analytics
    - Funding rate calculations and history
    - Open interest tracking
    - Insurance fund monitoring
    - Platform-wide statistics
    - WebSocket support for live updates
    
    ## Rate Limits
    - Read endpoints: 100 requests/minute
    - Write endpoints: 10 requests/minute
    
    ## Caching
    Most endpoints implement response caching with varying TTLs (10-60 seconds) to optimize performance.
    
  contact:
    name: Percolator Team
    url: https://percolator.trade
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Local development server
  - url: https://api.percolator.trade
    description: Production server
...
paths:
...
  /funding/global:
    get:
      tags:
        - Funding
      summary: Get current funding rates for all markets
      description: Returns current funding rate data for all markets (60s cache)
      operationId: getGlobalFunding
      responses:
        '200':
          description: Successfully retrieved global funding data
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  markets:
                    type: array
                    items:
                      $ref: '#/components/schemas/FundingRateSummary'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /funding/{slab}:
    get:
      tags:
        - Funding
      summary: Get funding rate for a market
      description: Returns current funding rate and 24h history for a specific market (30s cache)
      operationId: getMarketFunding
      parameters:
        - $ref: '#/components/parameters/SlabAddress'
      responses:
        '200':
          description: Successfully retrieved funding data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FundingRateDetails'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /funding/{slab}/history:
    get:
      tags:
        - Funding
      summary: Get funding rate history
      description: Returns historical funding rate data with optional time range
      operationId: getMarketFundingHistory
      parameters:
        - $ref: '#/components/parameters/SlabAddress'
        - name: limit
          in: query
          description: Maximum number of records to return (max 1000)
          schema:
            type: integer
            default: 100
            maximum: 1000
        - name: since
          in: query
          description: ISO timestamp to fetch records from
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Successfully retrieved funding history
          content:
            application/json:
              schema:
                type: object
                properties:
                  slabAddress:
                    type: string
                  count:
                    type: integer
                  history:
                    type: array
                    items:
                      $ref: '#/components/schemas/FundingHistoryEntry'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /funding/{slab}/historySince:
    get:
      tags:
        - Funding
      summary: Get funding history since a timestamp
      description: |
        Returns funding rate history starting from a required `since` timestamp.
        Similar to /funding/{slab}/history but with a mandatory start time.
      operationId: getMarketFundingHistorySince
      parameters:
        - $ref: '#/components/parameters/SlabAddress'
        - name: since
          in: query
          required: true
          description: |
            Start timestamp — ISO 8601 string or Unix epoch (seconds or milliseconds).
            Must resolve to a year between 2020 and 2100.
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of records to return (max 500)
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 500
      responses:
        '200':
          description: Successfully retrieved funding history
          content:
            application/json:
              schema:
                type: object
                properties:
                  slabAddress:
                    type: string
                  since:
                    type: string
                    format: date-time
                    description: Validated ISO 8601 start timestamp
                  count:
                    type: integer
                  history:
                    type: array
                    items:
                      $ref: '#/components/schemas/FundingHistoryEntry'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: API key passed via the x-api-key request header
  parameters:
    SlabAddress:
      name: slab
      in: path
      required: true
      description: Market slab address (Solana public key)
      schema:
        type: string
        pattern: '^[1-9A-HJ-NP-Za-km-z]{32,44}$'
  schemas:
...
    FundingRateSummary:
      type: object
      properties:
        slabAddress:
          type: string
        currentRateBpsPerSlot:
          type: number
          description: Funding rate in basis points per slot
        hourlyRatePercent:
          type: number
          description: Hourly funding rate as percentage
        dailyRatePercent:
          type: number
          description: Daily funding rate as percentage
        netLpPos:
          type: string
          description: Net LP position (inventory imbalance)
    FundingRateDetails:
      type: object
      properties:
        slabAddress:
          type: string
        currentRateBpsPerSlot:
          type: number
        hourlyRatePercent:
          type: number
        dailyRatePercent:
          type: number
        annualizedPercent:
          type: number
        netLpPos:
          type: string
        last24hHistory:
          type: array
          items:
            $ref: '#/components/schemas/FundingHistoryEntry'
        metadata:
          type: object
          properties:
            dataPoints24h:
              type: integer
            explanation:
              type: object
              properties:
                rateBpsPerSlot:
                  type: string
                hourly:
                  type: string
                daily:
                  type: string
                annualized:
                  type: string
                sign:
                  type: string
                inventory:
                  type: string
    FundingHistoryEntry:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        slot:
          type: string
        rateBpsPerSlot:
          type: number
        netLpPos:
          type: string
        priceE6:
          type: string
        fundingIndexQpbE6:
          type: string
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
        details:
          type: string
          description: Additional error details
        hint:
          type: string
          description: Hint for resolving the error
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Invalid slab address
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Market stats not found
            hint: Market may not have been cranked yet or does not exist
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Failed to fetch data
            details: Database connection timeout